+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2023-07-31
+
+### Changed
+- New authentication method that uses OpenID Connect (OIDC) instead Oauth to avoid using secrets
+
+### Removed
+- Removed a few classes related to code authentication
+
 ## [1.3.2] - 2023-07-26
 
 ### Changed